- Enable warning level 4 for all configurations. - Fixed warnings (level 4).

+++ source/codec/compressor.h
@@ -62,12 +62,12 @@ public:
     // useful for end of the compression stream.
     //
     virtual bool Process(const uint8_t* input_data,
-                         int input_size,
+                         size_t input_size,
                          uint8_t* output_data,
-                         int output_size,
+                         size_t output_size,
                          CompressorFlush flush,
-                         int* consumed,
-                         int* written) = 0;
+                         size_t* consumed,
+                         size_t* written) = 0;
 };
 
 } // namespace aspia
